refactor: extract queries and dtos to separated projects
