+++ Source/Iwentys.sln
@@ -72,7 +72,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Iwentys.AccountManagement.C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Iwentys.AccountManagement.Server", "AccountManagement\Iwentys.AccountManagement.Server\Iwentys.AccountManagement.Server.csproj", "{915C94CD-61F8-41A8-9F9A-2B2185D02986}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Iwentys.Modules.Assignments", "Modules\Iwentys.Modules.Assignments\Iwentys.Modules.Assignments.csproj", "{555A4C72-75C8-458F-B79D-B6D8C4304A8B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Iwentys.Modules.Assignments", "Modules\Iwentys.Modules.Assignments\Iwentys.Modules.Assignments.csproj", "{555A4C72-75C8-458F-B79D-B6D8C4304A8B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Iwentys.Modules.Newsfeeds", "Modules\Iwentys.Modules.Newsfeeds\Iwentys.Modules.Newsfeeds.csproj", "{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Iwentys.Modules.Companies", "Modules\Iwentys.Modules.Companies\Iwentys.Modules.Companies.csproj", "{5978A0BB-B706-4ED3-8889-D09C3F2087A3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -184,6 +188,14 @@ Global
 		{555A4C72-75C8-458F-B79D-B6D8C4304A8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{555A4C72-75C8-458F-B79D-B6D8C4304A8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{555A4C72-75C8-458F-B79D-B6D8C4304A8B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5978A0BB-B706-4ED3-8889-D09C3F2087A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5978A0BB-B706-4ED3-8889-D09C3F2087A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5978A0BB-B706-4ED3-8889-D09C3F2087A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5978A0BB-B706-4ED3-8889-D09C3F2087A3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -215,6 +227,8 @@ Global
 		{B5D2A51D-3562-4EB9-BFCF-A5991C00BBE2} = {500314E4-05C5-4A6E-A6CE-08194C3DCAD6}
 		{915C94CD-61F8-41A8-9F9A-2B2185D02986} = {500314E4-05C5-4A6E-A6CE-08194C3DCAD6}
 		{555A4C72-75C8-458F-B79D-B6D8C4304A8B} = {8F56593F-F2A4-4421-B91D-D8B2E84ED39E}
+		{1478A9F6-58D6-4AC7-B1B8-7809CCF10C23} = {8F56593F-F2A4-4421-B91D-D8B2E84ED39E}
+		{5978A0BB-B706-4ED3-8889-D09C3F2087A3} = {8F56593F-F2A4-4421-B91D-D8B2E84ED39E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C699E386-11E7-4ABB-9AD6-C676F3CA2061}
