+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/Queries/GetAllSubjectsMentors.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -7,10 +6,8 @@ using AutoMapper;
 using Iwentys.Common.Exceptions;
 using Iwentys.Domain.AccountManagement;
 using Iwentys.Domain.AccountManagement.Mentors.Dto;
-using Iwentys.Domain.Study;
 using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
-using LanguageExt;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
