+++ Source/Modules/Iwentys.Modules.AccountManagement/Mentors/Queries/GetMentorsByGroupSubjectId.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -9,7 +8,6 @@ using Iwentys.Domain.AccountManagement.Mentors.Dto;
 using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
 using MediatR;
-using Microsoft.EntityFrameworkCore;
 
 namespace Iwentys.Modules.AccountManagement.Mentors.Queries
 {
