+++ Source/Modules/Iwentys.Modules.Companies/Dtos/CompanyWorkRequestDto.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Linq.Expressions;
 using Iwentys.Domain.AccountManagement.Dto;
+using Iwentys.Domain.Companies;
 
-namespace Iwentys.Domain.Companies.Dto
+namespace Iwentys.Modules.Companies.Dtos
 {
     public record CompanyWorkRequestDto
     {
