+++ Source/Modules/Iwentys.Modules.Companies/Iwentys.Modules.Companies.csproj
@@ -1,11 +1,10 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Infrastructure\Iwentys.Infrastructure.Application\Iwentys.Infrastructure.Application.csproj" />
     <ProjectReference Include="..\Iwentys.Modules.AccountManagement\Iwentys.Modules.AccountManagement.csproj" />
   </ItemGroup>
