+++ Source/Modules/Iwentys.Modules.Gamification/Quests/Dtos/QuestRatingRow.cs
