+++ Source/Modules/Iwentys.Modules.Gamification/Quests/Dtos/QuestResponseInfoDto.cs
