+++ Source/Modules/Iwentys.Modules.Newsfeeds/Iwentys.Modules.Newsfeeds.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Infrastructure\Iwentys.Infrastructure.Application\Iwentys.Infrastructure.Application.csproj" />
-    <ProjectReference Include="..\Iwentys.Modules.AccountManagement\Iwentys.Modules.AccountManagement.csproj" />
   </ItemGroup>
 
 </Project>
