+++ Source/Modules/Iwentys.Modules.PeerReview/Queries/SendReviewFeedback.cs
@@ -4,10 +4,12 @@ using AutoMapper;
 using Iwentys.Domain.AccountManagement;
 using Iwentys.Domain.PeerReview;
 using Iwentys.Domain.PeerReview.Dto;
+using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Modules.PeerReview.Dtos;
 using MediatR;
 
-namespace Iwentys.Infrastructure.Application.Controllers.PeerReview
+namespace Iwentys.Modules.PeerReview.Queries
 {
     public class SendReviewFeedback
     {
